+++ RELEASE_NOTES.md
@@ -10,6 +10,7 @@
 ### Small changes
 - Updated dependencies
 - Updated composer to 2.8.3
+- Updated wording for source URL #115
 
 
 ## Version 2024-11