+++ RELEASE_NOTES.md
@@ -1,5 +1,22 @@
 # Release notes
 
+## Version 2024-12
+*Release date: 2024-12-11*   
+
+### Bug fix
+- Fix: Sorting of lists in login area does not work properly #116
+
+### Changes
+- Added unique validation for invite translation - language id #59
+- Added unique validation for faq questions - question #59
+- Cleaned up and removed default comments in dhcrcore module #none
+
+### Small changes
+- Updated dependencies
+- Updated composer to 2.8.4
+- Updated wording for source URL #115
+
+
 ## Version 2024-11
 *Release date: 2024-11-15*